+++ History.md
@@ -1,3 +1,12 @@
 # History
 
 ### Version 1.0
+Einfügen von zwei Funktionen, KTS_LiveCheck und KTS_UpdateCheck.
+Erstellen vom Konfigurationsformular mit folgenden Konfigurationsmöglichkeiten:
+    - SMTP Instanz für den Alive Check
+    - SMTP Instanz für den Aktuallisierungs Check
+    - Eingabe der Lizenz
+    - Liste mit der zu Prüfende Variablen
+    - Intervall Zeit in Minuten für den Alive Check
+    - Intervall Zeit in Minuten für den Aktuallisierungs Check
+    - Notwendige Differenzzeit (letzte Aktuallisierung bis jetzt) für senden einer E-Mail